Add counting helpers
